requirements.txt
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
Pillow==10.2.0
Flask==3.0.0
